Updatenew

+++ 27-FrontToBackSqlConnection/Areas/AdminPanel/Controllers/ProductController.cs
@@ -33,7 +33,7 @@ public class ProductController : Controller
                 CategoryName = p.Category.Name,
                 Price = p.Price,
                 SKU = p.SKU,
-                Image = p.ProductImages.FirstOrDefault().Image,
+                Image = p.ProductImages.FirstOrDefault(p=>p.IsPrimary==true).Image,
             })
             .ToListAsync();
 
@@ -147,7 +147,6 @@ public class ProductController : Controller
                     IsPrimary=null
                 });
             }
-
         }
         TempData["FileInfo"] = info;
         
@@ -188,15 +187,49 @@ public class ProductController : Controller
     {
         if (id == null || id < 1) return BadRequest();
         
+        Product? existProduct = await _context.Products
+            .Include(p=>p.ProductImages)
+            .Include(p=>p.ProductTags)
+            .FirstOrDefaultAsync(p => p.Id == id);
+        
+        if (existProduct == null) return NotFound();
+        
         productUpdateVM.Categories = await _context.Categories.Where(c => !c.isDeleted).ToListAsync();
         productUpdateVM.Tags = await _context.Tags.Where(t => !t.isDeleted).ToListAsync();
+        productUpdateVM.ProductImages = existProduct.ProductImages;
         
         if (!ModelState.IsValid) return View(productUpdateVM);
+
         
+        if (productUpdateVM.MainPhoto is not null)
+        {
+            if (!productUpdateVM.MainPhoto.CheckFileType("image/"))
+            {
+                ModelState.AddModelError(nameof(productUpdateVM.MainPhoto), "File type is incorrect");
+                return View(productUpdateVM);
+            }
+            if (!productUpdateVM.MainPhoto.CheckFileSize(FileSize.MB,1))
+            {
+                ModelState.AddModelError(nameof(productUpdateVM.MainPhoto), "File size  less be than 2mb");
+                return View(productUpdateVM);
+            }
+        }
 
-        Product? existProduct = await _context.Products.FirstOrDefaultAsync(p => p.Id == id);
-        if (existProduct == null) return NotFound();
-
+        if (productUpdateVM.HoverPhoto is not null)
+        {
+            if (!productUpdateVM.HoverPhoto.CheckFileType("image/"))
+            {
+                ModelState.AddModelError(nameof(productUpdateVM.HoverPhoto), "File type is incorrect");
+                return View(productUpdateVM);
+            }
+        
+            if (!productUpdateVM.HoverPhoto.CheckFileSize(FileSize.MB,1))
+            {
+                ModelState.AddModelError(nameof(productUpdateVM.HoverPhoto), "File size  less be than 2mb");
+                return View(productUpdateVM);
+            }
+        }
+        
         bool existCategory = productUpdateVM.Categories.Any(c => c.Id == productUpdateVM.CategoryId);
         if (!existCategory)
         {
@@ -218,19 +251,85 @@ public class ProductController : Controller
         {
             productUpdateVM.TagIds = new();
         }
-
-
-        _context.ProductTags.RemoveRange(existProduct.ProductTags
-            .Where(pTag => !productUpdateVM.TagIds
-                .Exists(tId => tId == pTag.TagId))
-            .ToList());
         
-        _context.ProductTags.AddRange(productUpdateVM.TagIds
-            .Where(tId => !existProduct.ProductTags.Exists(pTag => pTag.TagId == tId))
-            .ToList()
-            .Select(tId => new ProductTag { TagId = tId, ProductId = existProduct.Id }));
+        if (productUpdateVM.TagIds is not  null)
+        {
+            _context.ProductTags.RemoveRange(existProduct.ProductTags
+                .Where(pTag => !productUpdateVM.TagIds
+                    .Exists(tId => tId == pTag.TagId))
+                .ToList());
+        
+            _context.ProductTags.AddRange(productUpdateVM.TagIds
+                .Where(tId => !existProduct.ProductTags.Exists(pTag => pTag.TagId == tId))
+                .ToList()
+                .Select(tId => new ProductTag { TagId = tId, ProductId = existProduct.Id }));
+        }
 
-
+        if (productUpdateVM.MainPhoto is not null)
+        {
+            string fileName = await productUpdateVM.MainPhoto.CreateFile(_env.WebRootPath,"assets","images","website-images");
+            ProductImage mainImage = existProduct.ProductImages.FirstOrDefault(p => p.IsPrimary == true);
+            
+            mainImage.Image.DeleteFile(_env.WebRootPath,"assets","images","website-images");
+            existProduct.ProductImages.Remove(mainImage);
+            existProduct.ProductImages.Add(new ProductImage
+            {
+                Image = fileName,
+                IsPrimary = true
+            });
+        }
+        if (productUpdateVM.HoverPhoto is not null)
+        {
+            string fileName = await productUpdateVM.HoverPhoto.CreateFile(_env.WebRootPath,"assets","images","website-images");
+            ProductImage hoverImage = existProduct.ProductImages.FirstOrDefault(p => p.IsPrimary == false);
+            
+            hoverImage.Image.DeleteFile(_env.WebRootPath,"assets","images","website-images");
+            existProduct.ProductImages.Remove(hoverImage);
+            existProduct.ProductImages.Add(new ProductImage
+            {
+                Image = fileName,
+                IsPrimary = false
+            });
+        }
+        
+        if (productUpdateVM.ImageIds is not null)
+        {
+            productUpdateVM.ImageIds = new List<int>();
+        }
+        
+        var deleteImages =
+            existProduct.ProductImages
+                .Where(pi => productUpdateVM.ImageIds
+                    .Exists(imgId => imgId == pi.Id )&& pi.IsPrimary==true)
+                .ToList();
+        deleteImages.ForEach(di=>di.Image.DeleteFile(_env.WebRootPath,"assets","images","website-images"));
+        _context.ProductImages.RemoveRange(deleteImages);
+        
+        if (productUpdateVM.AdditionalPhotos is not null)
+        {
+            string info = string.Empty;
+            foreach (var file in productUpdateVM.AdditionalPhotos)
+            {
+                if (!file.CheckFileType("image/"))
+                {
+                    info += $"<p>{file.FileName} type was not correct</p>";
+                    continue;
+                }
+                if (!file.CheckFileSize(FileSize.KB,100))
+                {
+                    info += $"<p>{file.FileName} size was not correct</p>";
+                    continue;
+                }
+            
+                existProduct.ProductImages.Add(new ProductImage
+                {
+                    Image = await file.CreateFile(_env.WebRootPath, "assets", "images", "website-images"),
+                    IsPrimary=null
+                });
+            }
+            TempData["FileInfo"] = info;
+        }
+        
         existProduct.Name = productUpdateVM.Name;
         existProduct.Price = productUpdateVM.Price;
         existProduct.Description = productUpdateVM.Description;
